requirements.txt
flask>=2.2.0
sentence-transformers>=2.2.2
transformers
torch>=1.13.0
google-search-results
faiss-cpu>=1.7.4
scikit-learn>=1.2.2
numpy>=1.24.0
pandas>=2.0.0
tqdm
openai>=0.27.0
requests